backend/requirements.txt
Django==4.2.7
qrcode==7.4.2
Pillow==10.1.0
django-cors-headers==4.3.1

